+++ CHANGELOG.md
@@ -44,11 +44,15 @@
 ### Alterado
 *  Argumento "python -m" adicionado antes do "pip install" no Dockerfile
 
+## [0.2.0] - 2019-11-02
+### Alterado
+* Mudança na organização dos arquivos do schema e do teste de accounts.
+
 ## [0.2.0] - 2019-11-03
 ### Adicionado
-*  Nova mutation update da model EntryVisitor.
-*  Nova mutation delete da model EntryVisitor.
+*  Nova mutation update e delete da model EntryVisitor.
 
 
  ---
+ 
  Modelo padrão do changelog disponível [aqui](https://keepachangelog.com/en/0.3.0/).